Implement Assignment 4: operator<<m virtual methods, 9-object vectorm full tests
