--- Assignment3/Alliance.cpp
@@ -1,17 +1,0 @@
-/*
- * Ваганов Владимир Алексеевич
- * st140060@student.spbu.ru
- * Assignment 4
- */
-
-#include "Alliance.h"
-
-Alliance::Alliance() : name_("Unknown Alliance") {}
-
-Alliance::Alliance(const std::string& name)
-    : name_(name) {}
-
-const std::string& Alliance::getName() const
-{
-    return name_;
-}
