--- Assignment3/Transformer.h
@@ -1,59 +1,0 @@
-/*
- * Ваганов Владимир Алексеевич
- * st140060@student.spbu.ru
- * Assignment 4
- */
-
-#ifndef TRANSFORMER_H
-#define TRANSFORMER_H
-
-#include "Gun.h"
-#include "Alliance.h"
-#include <string>
-#include <iostream>
-
-class Transformer
-{
-public:
-    Transformer();
-    Transformer(
-        const std::string& name,
-        int level,
-        int strength,
-        int fuel,
-        int ammo,
-        const std::string& gunModel,
-        Alliance* ally
-    );
-    virtual ~Transformer();
-
-    const std::string& getName() const;
-    int getLevel() const;
-    int getStrength() const;
-    int getFuel() const;
-    Gun& getGun();
-    Alliance* getAlliance() const;
-
-    void setLevel(int level);
-    void setStrength(int strength);
-    void setFuel(int fuel);
-
-    virtual bool move();
-    virtual bool fire();
-    virtual bool transform() = 0;
-    virtual void specialAbility() = 0;
-    virtual void print(std::ostream& os) const = 0;
-
-    friend std::ostream& operator<<(std::ostream& os, const Transformer& t);
-protected:
-    std::string name_;
-    int level_;
-    int strength_;
-    int fuel_;
-    Gun gun_;
-    Alliance* ally_;
-};
-
-std::ostream& operator<<(std::ostream& os, const Transformer& t);
-
-#endif
