+++ Assignment3/src/Alliance.cpp
@@ -1,7 +1,7 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #include "Alliance.h"
@@ -11,5 +11,8 @@ Alliance::Alliance() : name_("Unknown") {}
 Alliance::Alliance(const std::string& name)
     : name_(name) {}
 
-const std::string& Alliance::getName() const { return name_; }
+const std::string& Alliance::getName() const
+{
+    return name_;
+}
