+++ Assignment3/src/Autobot.h
@@ -1,15 +1,17 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #ifndef AUTOBOT_H
 #define AUTOBOT_H
 
 #include "Transformer.h"
+#include <iostream>
 
-class Autobot : public Transformer {
+class Autobot : public Transformer
+{
 public:
     Autobot();
     Autobot(
@@ -30,6 +32,8 @@ public:
 
     bool hasMatrixOfLeadership() const;
     void setMatrixOfLeadership(bool has);
+
+    void print(std::ostream& os) const override;
 
 private:
     bool hasMatrixOfLeadership_;
