+++ Assignment3/src/Decepticon.cpp
@@ -1,10 +1,11 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #include "Decepticon.h"
+#include <iostream>
 
 Decepticon::Decepticon()
     : Transformer("Decepticon", 1, 10, 10, 5, "Plasma", nullptr),
@@ -20,10 +21,42 @@ Decepticon::Decepticon(
     : Transformer(name, level, strength, fuel, ammo, gunModel, ally),
       hasDarkSpark_(hasDarkSpark) {}
 
-bool Decepticon::transform() { return true; }
-bool Decepticon::specialAbility() { return true; }
-bool Decepticon::activateStealth() { return true; }
-void Decepticon::corruptCore() {}
+bool Decepticon::transform()
+{
+    std::cout << "[Decepticon::transform] called" << std::endl;
+    return true;
+}
 
-bool Decepticon::hasDarkSpark() const { return hasDarkSpark_; }
-void Decepticon::setDarkSpark(bool has) { hasDarkSpark_ = has; }
+bool Decepticon::specialAbility()
+{
+    std::cout << "[Decepticon::specialAbility] called" << std::endl;
+    return true;
+}
+
+bool Decepticon::activateStealth()
+{
+    std::cout << "[Decepticon::activateStealth] called" << std::endl;
+    return true;
+}
+
+void Decepticon::corruptCore()
+{
+    std::cout << "[Decepticon::corruptCore] called" << std::endl;
+}
+
+bool Decepticon::hasDarkSpark() const
+{
+    return hasDarkSpark_;
+}
+void Decepticon::setDarkSpark(bool has)
+{
+    hasDarkSpark_ = has;
+}
+
+void Decepticon::print(std::ostream& os) const
+{
+    os << "[Decepticon] " << name_
+       << ", Level: " << level_
+       << ", Fuel: " << fuel_
+       << ", Has Dark Spark: " << (hasDarkSpark_ ? "yes" : "no");
+}
