+++ Assignment3/src/Gun.h
@@ -1,7 +1,7 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #ifndef GUN_H
@@ -9,7 +9,8 @@
 
 #include <string>
 
-class Gun {
+class Gun
+{
 public:
     Gun();
     Gun(const std::string& model, int ammo);
