+++ Assignment3/src/Pretender.cpp
@@ -1,10 +1,11 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #include "Pretender.h"
+#include <iostream>
 
 Pretender::Pretender()
     : Transformer("Pretender", 1, 10, 10, 5, "Shell", nullptr),
@@ -20,10 +21,42 @@ Pretender::Pretender(
     : Transformer(name, level, strength, fuel, ammo, gunModel, ally),
       isHumanShell_(isHuman) {}
 
-bool Pretender::transform() { return true; }
-bool Pretender::specialAbility() { return true; }
-bool Pretender::detachShell() { return true; }
-void Pretender::mimicAlliance() {}
+bool Pretender::transform()
+{
+    std::cout << "[Pretender::transform] called" << std::endl;
+    return true;
+}
 
-bool Pretender::isHumanShell() const { return isHumanShell_; }
-void Pretender::setHumanShell(bool is) { isHumanShell_ = is; }
+bool Pretender::specialAbility()
+{
+    std::cout << "[Pretender::specialAbility] called" << std::endl;
+    return true;
+}
+
+bool Pretender::detachShell()
+{
+    std::cout << "[Pretender::detachShell] called" << std::endl;
+    return true;
+}
+
+void Pretender::mimicAlliance()
+{
+    std::cout << "[Pretender::mimicAlliance] called" << std::endl;
+}
+
+bool Pretender::isHumanShell() const
+{
+    return isHumanShell_;
+}
+void Pretender::setHumanShell(bool is)
+{
+    isHumanShell_ = is;
+}
+
+void Pretender::print(std::ostream& os) const
+{
+    os << "[Pretender] " << name_
+       << ", Level: " << level_
+       << ", Fuel: " << fuel_
+       << ", Human Shell: " << (isHumanShell_ ? "yes" : "no");
+}
