+++ Assignment3/src/Transformer.cpp
@@ -1,10 +1,11 @@
 /*
  * Ваганов Владимир Алексеевич
  * st140060@student.spbu.ru
- * Assignment 3
+ * Assignment 4
  */
 
 #include "Transformer.h"
+#include <iostream>
 
 Transformer::Transformer()
     : name_("Unknown"), level_(1), strength_(10), fuel_(10),
@@ -21,17 +22,62 @@ Transformer::Transformer(
 
 Transformer::~Transformer() = default;
 
-const std::string& Transformer::getName() const { return name_; }
-int Transformer::getLevel() const { return level_; }
-int Transformer::getStrength() const { return strength_; }
-int Transformer::getFuel() const { return fuel_; }
-Gun& Transformer::getGun() { return gun_; }
-const Gun& Transformer::getGun() const { return gun_; }
-Alliance* Transformer::getAlliance() const { return ally_; }
+const std::string& Transformer::getName() const
+{
+    return name_;
+}
+int Transformer::getLevel() const
+{
+    return level_;
+}
+int Transformer::getStrength() const
+{
+    return strength_;
+}
+int Transformer::getFuel() const
+{
+    return fuel_;
+}
+Gun& Transformer::getGun()
+{
+    return gun_;
+}
+const Gun& Transformer::getGun() const
+{
+    return gun_;
+}
+Alliance* Transformer::getAlliance() const
+{
+    return ally_;
+}
 
-void Transformer::setLevel(int level) { level_ = level; }
-void Transformer::setStrength(int strength) { strength_ = strength; }
-void Transformer::setFuel(int fuel) { fuel_ = fuel; }
+void Transformer::setLevel(int level)
+{
+    level_ = level;
+}
+void Transformer::setStrength(int strength)
+{
+    strength_ = strength;
+}
+void Transformer::setFuel(int fuel)
+{
+    fuel_ = fuel;
+}
 
-bool Transformer::move() { return true; }
-bool Transformer::fire() { return true; }
+bool Transformer::move()
+{
+    std::cout << "[" << name_ << "::move] called" << std::endl;
+    return true;
+}
+
+bool Transformer::fire()
+{
+    std::cout << "[" << name_ << "::fire] called" << std::endl;
+    return true;
+}
+
+std::ostream& operator<<(std::ostream& os, const Transformer& t)
+{
+    t.print(os);
+    return os;
+}
